feat: move methods for work with photo from file package to photo package
